Reorganize slots
This moves the code for showing new metadata into the state change
slot. Metadata will be known after the file is loaded. This prevents
this code from breaking if the file has no metadata.

This also moves some code into a source changed slot. It's a bit cleaner
this way.